# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "operationId": "JobsController_list_v1",
        "parameters": [
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "description": "Sort direction.",
            "schema": {
              "$ref": "#/components/schemas/SortDir"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort column.",
            "schema": {
              "$ref": "#/components/schemas/JobSortBy"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Page offset.",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Full-text query over title + company + description.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "End of date interval (ISO 8601).",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Start of date interval (ISO 8601).",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dateField",
            "required": false,
            "in": "query",
            "description": "Date field for interval filtering.",
            "schema": {
              "$ref": "#/components/schemas/DateField"
            }
          },
          {
            "name": "salaryMax",
            "required": false,
            "in": "query",
            "description": "Maximum salary.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "salaryMin",
            "required": false,
            "in": "query",
            "description": "Minimum salary.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scoreMax",
            "required": false,
            "in": "query",
            "description": "Maximum match score (0–100).",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scoreMin",
            "required": false,
            "in": "query",
            "description": "Minimum match score (0–100).",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "reaction",
            "required": false,
            "in": "query",
            "description": "Comma-separated reaction stage values.",
            "schema": {
              "example": "saved,applied",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Comma-separated status values.",
            "schema": {
              "example": "new,processed",
              "type": "string"
            }
          },
          {
            "name": "seniority",
            "required": false,
            "in": "query",
            "description": "Comma-separated seniority values.",
            "schema": {
              "example": "senior,lead",
              "type": "string"
            }
          },
          {
            "name": "remote",
            "required": false,
            "in": "query",
            "description": "Comma-separated remote values.",
            "schema": {
              "example": "remote,hybrid",
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tags.",
            "schema": {
              "example": "typescript,node",
              "type": "string"
            }
          },
          {
            "name": "sources",
            "required": false,
            "in": "query",
            "description": "Comma-separated source ids.",
            "schema": {
              "example": "1,3",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedJobsResponse"
                }
              }
            }
          }
        },
        "summary": "List jobs with filters, full-text search, and pagination",
        "tags": ["jobs"]
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "operationId": "JobsController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job id (bigint as string).",
            "schema": {
              "example": "42"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found."
          }
        },
        "summary": "Get a single job by id",
        "tags": ["jobs"]
      }
    },
    "/v1/jobs/{id}/status": {
      "patch": {
        "operationId": "JobsController_setStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job id (bigint as string).",
            "schema": {
              "example": "42"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetJobStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          }
        },
        "summary": "Update the status of a job (archive, hide, restore)",
        "tags": ["jobs"]
      }
    },
...
  },
  "info": {
    "title": "job-hunter API",
    "description": "Gateway for jobs, reactions, keyword dictionaries, profiles, LLM admin, and sources.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SortDir": {
        "type": "string",
        "enum": ["asc", "desc"]
      },
      "JobSortBy": {
        "type": "string",
        "enum": ["score", "posted", "salary", "lastSeen", "board"]
      },
      "DateField": {
        "type": "string",
        "enum": ["posted", "first_seen"]
      },
      "Seniority": {
        "type": "string",
        "enum": ["junior", "middle", "senior", "lead", "unknown"],
        "description": "Seniority level."
      },
      "RemoteType": {
        "type": "string",
        "enum": ["remote", "hybrid", "office", "unknown"],
        "description": "Remote arrangement."
      },
      "JobStatus": {
        "type": "string",
        "enum": ["new", "processed", "archived", "hidden"],
        "description": "Job status."
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Primary key (bigint serialized as a string).",
            "example": "42"
          },
          "sourceId": {
            "type": "number",
            "description": "Source identifier."
          },
          "sourceSlug": {
            "type": "string",
            "description": "Source slug (denormalized for display).",
            "example": "hh"
          },
          "externalId": {
            "type": "string",
            "description": "External posting id on the source site."
          },
          "url": {
            "type": "string",
            "description": "Canonical URL."
          },
          "title": {
            "type": "string",
            "description": "Job title."
          },
          "company": {
            "type": "string",
            "description": "Company name, when known.",
            "nullable": true
          },
          "descriptionMd": {
            "type": "string",
            "description": "Job description in markdown.",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "description": "LLM-generated summary.",
            "nullable": true
          },
          "tags": {
            "description": "Extracted tags.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "redFlags": {
            "description": "Detected red flags.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "salaryMin": {
            "type": "number",
            "description": "Minimum salary, when posted.",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "description": "Maximum salary, when posted.",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "description": "Salary currency code.",
            "nullable": true
          },
          "seniority": {
            "description": "Seniority level.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              }
            ]
          },
          "remote": {
            "description": "Remote arrangement.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RemoteType"
              }
            ]
          },
          "location": {
            "type": "string",
            "description": "Location, when known.",
            "nullable": true
          },
          "postedAt": {
            "type": "string",
            "description": "Posting date on the source site (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "firstSeenAt": {
            "type": "string",
            "description": "When the job was first seen by a scraper (ISO 8601).",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "description": "When the job was last seen by a scraper (ISO 8601).",
            "format": "date-time"
          },
          "status": {
            "description": "Job status.",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              }
            ]
          },
          "matchScore": {
            "type": "number",
            "description": "Latest match score for the active profile, if any.",
            "nullable": true
          },
          "currentReaction": {
            "type": "string",
            "description": "Current reaction stage for the active profile, if any.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sourceId",
          "sourceSlug",
          "externalId",
          "url",
          "title",
          "company",
          "descriptionMd",
          "summary",
          "tags",
          "redFlags",
          "salaryMin",
          "salaryMax",
          "salaryCurrency",
          "seniority",
          "remote",
          "location",
          "postedAt",
          "firstSeenAt",
          "lastSeenAt",
          "status",
          "matchScore",
          "currentReaction"
        ]
      },
      "PaginatedJobsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Items on the current page.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of items matching the filter."
          }
        },
        "required": ["items", "total"]
      },
      "JobDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Primary key (bigint serialized as a string).",
            "example": "42"
          },
          "sourceId": {
            "type": "number",
            "description": "Source identifier."
          },
          "sourceSlug": {
            "type": "string",
            "description": "Source slug (denormalized for display).",
            "example": "hh"
          },
          "externalId": {
            "type": "string",
            "description": "External posting id on the source site."
          },
          "url": {
            "type": "string",
            "description": "Canonical URL."
          },
          "title": {
            "type": "string",
            "description": "Job title."
          },
          "company": {
            "type": "string",
            "description": "Company name, when known.",
            "nullable": true
          },
          "descriptionMd": {
            "type": "string",
            "description": "Job description in markdown.",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "description": "LLM-generated summary.",
            "nullable": true
          },
          "tags": {
            "description": "Extracted tags.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "redFlags": {
            "description": "Detected red flags.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "salaryMin": {
            "type": "number",
            "description": "Minimum salary, when posted.",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "description": "Maximum salary, when posted.",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "description": "Salary currency code.",
            "nullable": true
          },
          "seniority": {
            "description": "Seniority level.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              }
            ]
          },
          "remote": {
            "description": "Remote arrangement.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RemoteType"
              }
            ]
          },
          "location": {
            "type": "string",
            "description": "Location, when known.",
            "nullable": true
          },
          "postedAt": {
            "type": "string",
            "description": "Posting date on the source site (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "firstSeenAt": {
            "type": "string",
            "description": "When the job was first seen by a scraper (ISO 8601).",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "description": "When the job was last seen by a scraper (ISO 8601).",
            "format": "date-time"
          },
          "status": {
            "description": "Job status.",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              }
            ]
          },
          "matchScore": {
            "type": "number",
            "description": "Latest match score for the active profile, if any.",
            "nullable": true
          },
          "currentReaction": {
            "type": "string",
            "description": "Current reaction stage for the active profile, if any.",
            "nullable": true
          },
          "matchExplanation": {
            "type": "string",
            "description": "LLM rationale behind matchScore, read from job_matches.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sourceId",
          "sourceSlug",
          "externalId",
          "url",
          "title",
          "company",
          "descriptionMd",
          "summary",
          "tags",
          "redFlags",
          "salaryMin",
          "salaryMax",
          "salaryCurrency",
          "seniority",
          "remote",
          "location",
          "postedAt",
          "firstSeenAt",
          "lastSeenAt",
          "status",
          "matchScore",
          "currentReaction",
          "matchExplanation"
        ]
      },
      "SetJobStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "New job status.",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              }
            ]
          }
        },
        "required": ["status"]
      },
...
    }
  }
}